Refactoring done in tests and core framework.
