+++ src/main/java/com/github/wasiqb/coteafs/appium/config/ios/IOSArgumentSetting.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.wasiqb.coteafs.appium.config;
+package com.github.wasiqb.coteafs.appium.config.ios;
 
 /**
  * @author wasiq.bhamla
