+++ src/test/resources/appium-config.yaml
@@ -27,23 +27,28 @@ device: &default_device
   device_type: REAL
   session_timeout: 120000
   clear_system_files: true
-#  no_reset: false
-#  full_reset: true
+  no_reset: false
+  full_reset: true
   playback:
     record:
       enabled: false
     delay_before_swipe: 200
     delay_after_swipe: 100
-    delay_before_tap: 200
-    delay_after_tap: 100
+    delay_before_tap: 0
+    delay_after_tap: 0
 
 devices:
   test_local:
-    device_name: MI Redmi Note 4
-    device_version: 7.0
+    device_name: pixel2
+    device_version: 8.1
     app_location: apps/android/VodQA.apk
-    automation_name: UIAUTOMATOR2
+    automation_name: APPIUM
     <<: *default_device
+    device_type: SIMULATOR
+    android:
+      avd: pixel2
+      avd_launch_timeout: 60000
+      avd_ready_timeout: 60000
   test_browserstack:
     device_name: Samsung Galaxy S8
     device_version: 7.0
